requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-login
python-dotenv
werkzeug
numpy
pandas
scikit-learn
joblib
